+++ wingle/src/main/java/kr/co/wingle/member/AuthController.java
@@ -19,6 +19,8 @@ import kr.co.wingle.member.dto.SignupRequestDto;
 import kr.co.wingle.member.dto.SignupResponseDto;
 import kr.co.wingle.member.dto.TokenDto;
 import kr.co.wingle.member.dto.TokenRequestDto;
+import kr.co.wingle.member.entity.Member;
+import kr.co.wingle.member.service.AuthService;
 import lombok.RequiredArgsConstructor;
 
 @RestController
