+++ wingle/src/main/java/kr/co/wingle/member/MemberController.java
@@ -2,6 +2,7 @@ package kr.co.wingle.member;
 
 import org.springframework.web.bind.annotation.RestController;
 
+import kr.co.wingle.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 
 @RestController
