+++ wingle/src/main/java/kr/co/wingle/member/MemberRepository.java
@@ -4,6 +4,8 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import kr.co.wingle.member.entity.Member;
+
 public interface MemberRepository extends JpaRepository<Member, Long> {
 	boolean existsByEmail(String email);
