+++ wingle/src/main/java/kr/co/wingle/member/entity/Authority.java
@@ -1,4 +1,4 @@
-package kr.co.wingle.member;
+package kr.co.wingle.member.entity;
 
 public enum Authority {
 	ROLE_USER, ROLE_ADMIN;
