+++ wingle/src/main/java/kr/co/wingle/member/entity/Permission.java
@@ -1,4 +1,4 @@
-package kr.co.wingle.member;
+package kr.co.wingle.member.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
