+++ wingle/src/test/java/kr/co/wingle/member/AuthServiceTest.java
@@ -18,6 +18,8 @@ import kr.co.wingle.member.dto.SignupRequestDto;
 import kr.co.wingle.member.dto.SignupResponseDto;
 import kr.co.wingle.member.dto.TokenDto;
 import kr.co.wingle.member.dto.TokenRequestDto;
+import kr.co.wingle.member.entity.Member;
+import kr.co.wingle.member.service.AuthService;
 
 @SpringBootTest
 @WithMockUser(value = EMAIL, password = PASSWORD)
